теперь все переменные по умолчанию имеют тип вещественных чисел, соответственно обновил арифметику чтобы она могла считать дроби

+++ app/src/main/java/com/example/codeblocks/model/Interpretator.kt
@@ -2,7 +2,7 @@ package com.example.codeblocks.model
 
 object Interpretator {
     // словарь переменных, которые создаются по ходу действия программы
-    private val variables: MutableMap<String, Int> = mutableMapOf()
+    private val variables: MutableMap<String, Double> = mutableMapOf()
 
     // выполнить код
     fun run(code: MutableList<Command>) {
